--- diskann-quantization/src/multi_vector/distance_v2/kernel/panel/maxsim/mod.rs
@@ -1,6 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-pub(crate) mod f32;
