--- diskann-quantization/src/multi_vector/distance_v2/kernel/panel/mod.rs
@@ -1,10 +1,0 @@
-/*
- * Copyright (c) Microsoft Corporation.
- * Licensed under the MIT license.
- */
-
-pub(crate) mod maxsim;
-
-pub(crate) trait Kernel {
-    fn run(&mut self);
-}
