+++ diskann-quantization/src/multi_vector/distance_v2/mod.rs
@@ -12,4 +12,3 @@ mod util;
 
 pub mod tile;
 
-use bounds::{Check, Length};
